refactor algorithim components
